requirements.txt
psycopg2-binary==2.9.1
Flask~=2.0.2
requests~=2.26.0